Create cancel-edit function & modify value getting
Instead of performing the query selecting from
within the submission handler, create function
getValues() that queries the form based on each
input's name attribute. Using the name attribute
allows to create a more generic function that can
grab values from a form. However, if both sets of
inputs have the same name attribute, there could
be an issue with grabbing the wrong input if
both forms (the modal and edit pane) are open at
the same time. Therefore, take an argument in
getValues that will be the correct form and query
with the form's className -- then the name
attribute. This ensures the correct values being
grabbed no matter which form calls the function
or if both are open.