Modify mutation observer that closes editPane
In modalController, append confirm modals to #items
instead, allowing for the mutationObserver to
detect them. Give them each a unique class so
the mutationObserver can detect if it is the
modal that is responsible for deleting projects.
This way, the edit pane will close when any modals
are opened, except for the confirm-cancel-edit
modal.