Fix tasks not displaying properly with filters
Tasks were only being refreshed if a project filter
was open, not if a filter like "today" was open.
Modify displayController.currentOpenProject to
currentOpenFilter, change filterTasks to accept
a third optional parameter which will be the
button corresponding to the active filter. If not
passed at all, set displayController.#activeFilter
to null once again from filterTasks. If the filter
is active, "click" the stored button to refresh
the page with the appropriate tasks on every
new task creation.