Allow for taskGroup functionality
With the filters of "Today" and "Upcoming", the
desired behavior is to show relevant tasks as well
as overdue tasks. In order to do this, create
a TaskGroup class which stores the relevant tasks
in a container with a header and dropdown
corresponding to its group. This somewhat broke
the task adding/removing/editPane functionality,
so change the DOM removing/adding of both tasks
and editPanes to access their parentNodes in
relevant functions, rather than
displayController.#items. Also, this broke the
mutationObserver of the editPane, requiring instead
an event listener to the window that listens for
a click anywhere outside of the editPane, closing
it immediately and inserting the previously
existing task.