Fix sidebar bug & change when taskGroups are made
The overlay listener on the sidebar was firing
during the close animation of the sidebar if you
clicked on the open/close sidebar button and then
clicked quickly a second time, causing the behavior
to break. Set once to true in the event listener
to fix this. Also, only create task groups when
there are actually overdue tasks. Otherwise, display
them as is.