Change editPane listener to 'click'
The event listener attached to a task's editPane
was set to listen for a 'mousedown' anywhere
outside of the editPane, which would cause it to
close. However, the type being 'mousedown' caused
a bug: If an editPane was already open and the user
clicked the edit button of a task after it
(not before), the editPane would close but the
next editPane that the user tried to open would
not open. Change the event type to 'click' and
set a delay for the attaching of the listener so
that it doesn't immediately fire and close the edit.